+++ get_com_select_insert.h
@@ -6,12 +6,12 @@
 /**
 Reads COM_SELECT and COM_INSERT variables from all nodes and stores into 'selects' and 'inserts'
 */
-int get_global_status_allnodes(int *selects, int *inserts, Mariadb_nodes * nodes, int silent);
+int get_global_status_allnodes(long int *selects, long int *inserts, Mariadb_nodes * nodes, int silent);
 
 /**
 Prints difference in COM_SELECT and COM_INSERT 
 */
-int print_delta(int *new_selects, int *new_inserts, int *selects, int *inserts, int NodesNum);
+int print_delta(long int *new_selects, long int *new_inserts, long int *selects, long int *inserts, int NodesNum);
 
 
 #endif // GET_COM_SELECT_INSERT_H
